Update Sphinx to 3.x to see if that fixes the PDF build on Read the Docs.
No significant changes noticed in the HTML output compared to 1.8.5 (previously used by RTD) and in the PDF output compared to 2.4.4 (previously used by me locally).

Not updating all the way to the current version 4.x because that uses a different monospaced font in the PDF output, which disrupts the handout layout. Will try to fix that next.